script/requirements.txt
mysql-connector-python
ntscraper
pandas 
pyarrow
tqdm